Fix `retry-after` value in distributed case
The basic approach in #30 -- to sync around the timestamp of the oldest
event -- is sound, but the implementation was flawed, because the spot
before the cursor is not always the oldest event in the ring buffer. We
now correctly compute that value while counting events in the window
(which we had to do in order to sync event counts anyway). Additionally,
this commit adds tests for the distributed rate limiter which simulate a
peer by constructing a `ringBufferRateLimiter`, writing it out to
storage, and then starting up a `caddytest.Tester` on that storage.