Fix race when initializing ringBufferRateLimiter (#43)
This closes https://github.com/mholt/caddy-ratelimit/issues/36.

Fixes a race condition between ringBufferRateLimiter creation and its insertion
into a map. Do this by locking the entire map when we get or insert a
ringBufferRateLimiter.

I have replaced use of sync.Map with a normal  `map[string]*ringBufferRateLimiter`
and a `sync.Mutex`. They are passed around with a `rateLimitersMap` struct.
I've factored out logic into methods of rateLimitersMap, which enables some
careful use of defer `rlm.mu.Unlock()`` to avoid leaving a lock held open
on `panic()`.

We didn't see a need for a sync.Map. The docs suggest against using it for
type safety, and none of the suggested use cases apply. https://pkg.go.dev/sync#Map.
Let me know if I'm misunderstanding the use case (very possible!).

I've removed the sync.Pool, for now. Since ringBufferRateLimiter creation
and insertion is fully synchronized, I didn't see a need for it.

Note that some of the defensive refactoring is not strictly required--I have
a change that preserves the existing data structures, but I think the
suggested changeset is an overall improvement in maintainability.
https://github.com/divviup/caddy-ratelimit/pull/1/commits/65ad951ea012a5410dff297efa9da6f769e20dc0.

Some discussion of the performance impact and profiles is here
https://github.com/divviup/caddy-ratelimit/pull/1.
TL;DR, no meaningful impact to CPU, memory, or latency. This implementation could
be optimized by replacing the normal mutex with a RWMutex, but it would be
a marginal improvement (if any) in exchange for much more complicated locking
semantics.